requirements.txt
# Application dependencies
flask==3.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0

# Development dependencies
pytest==7.4.4
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pytest-cov==4.1.0 